Change how user is created
In this new implementation the user for the docker container is
created during image build and not at container initialization.

The user is a sudo user (with no password), alleviating permissions
problems from before.

This also changes how the images are used: it no longer makes sense
to download them from the repository, instead the user should
build them locally first and then use them. The process is explained
in the readme and will be simplified with an installation script
in the future